Add new request classes for managing cities and governorates; update validation rules to use correct columns; enhance `Bhry98CountriesResource` with filters and actions; optimize autoload configuration and update dependencies.

+++ src/Services/locations/CountriesManagementService.php
@@ -6,12 +6,10 @@ use Bhry98\Bhry98LaravelReady\Models\locations\LocationsCitiesModel;
 use Bhry98\Bhry98LaravelReady\Models\locations\LocationsCountriesModel;
 use Bhry98\Bhry98LaravelReady\Models\locations\LocationsGovernoratesModel;
 use Bhry98\Bhry98LaravelReady\Services\BaseService;
-use Filament\Notifications\Notification;
 use Illuminate\Pagination\LengthAwarePaginator;
 
 class CountriesManagementService extends BaseService
 {
-    // For Filament
     /**
      * @param int $id
      * @param array|null $relations
@@ -21,7 +19,23 @@ class CountriesManagementService extends BaseService
      */
     public function getById(int $id, array|null $relations = null, bool $withRelationsCount = false, bool $withTrash = false): ?LocationsCountriesModel
     {
-        $record = LocationsCountriesModel::query()->where(['id', $id]);
+        $record = LocationsCountriesModel::query()->where(['id' => $id]);
+        if ($withTrash) $record->withTrashed();
+        if ($relations) $record->with($relations);
+        if ($withRelationsCount) $record->withCount(['governorates', 'cities', 'users']);
+        return $record->first();
+    }
+
+    /**
+     * @param string $code
+     * @param array|null $relations
+     * @param bool $withRelationsCount
+     * @param bool $withTrash
+     * @return LocationsCountriesModel|null
+     */
+    public function getByCode(string $code, array|null $relations = null, bool $withRelationsCount = false, bool $withTrash = false): ?LocationsCountriesModel
+    {
+        $record = LocationsCountriesModel::query()->where(['code' => $code]);
         if ($withTrash) $record->withTrashed();
         if ($relations) $record->with($relations);
         if ($withRelationsCount) $record->withCount(['governorates', 'cities', 'users']);
@@ -35,6 +49,12 @@ class CountriesManagementService extends BaseService
     public function createCountry(array $data): bool
     {
         $record = LocationsCountriesModel::query()->create($data);
+        if (!$record) {
+            bhry98_send_filament_notification("danger", __("Bhry98::notifications.filament.created-field"));
+            return false;
+        } else {
+            bhry98_send_filament_notification("success", __("Bhry98::notifications.filament.created-success"));
+        }
         bhry98_created_log(success: (bool)$record, message: "create new country", context: ['record' => $record->toArray()]);
         return (bool)$record;
     }
@@ -47,22 +67,23 @@ class CountriesManagementService extends BaseService
     public function updateCountry(int $id, array $data): bool
     {
         $record = self::getById($id);
-        dd($id, $record);
         if (!$record) {
-            Notification::make()
-                ->title(__("Bhry98::notifications.filament.updated-field"))
-                ->danger()
-                ->send();
+            bhry98_send_filament_notification("danger", __("Bhry98::notifications.filament.updated-field"));
             return false;
         }
-        $updated = $record?->update(collect($data)->except(['names'])?->toArray() ?? []);
+        $updated = $record->update(collect($data)->except(['names'])?->toArray() ?? []);
         if (array_key_exists('names', $data)) {
             foreach ($data['names'] as $locale => $value) {
-                $record?->setLocalized('name', $value, $locale);
+                $record->setLocalized('name', $value, $locale);
             }
         }
-        bhry98_updated_log(success: (bool)$updated, message: "update country", context: ['record' => $record?->toArray(), 'data' => $data]);
-        return (bool)$updated;
+        if ($updated) {
+            bhry98_send_filament_notification("success", __("Bhry98::notifications.filament.updated-success"));
+        } else {
+            bhry98_send_filament_notification("danger", __("Bhry98::notifications.filament.updated-failed"));
+        }
+        bhry98_updated_log(success: $updated, message: "update country", context: ['record' => $record?->toArray(), 'data' => $data]);
+        return $updated;
     }
 
     /**
@@ -70,18 +91,34 @@ class CountriesManagementService extends BaseService
      * @param bool $force
      * @return bool
      */
-    public function deleteCountry(int $id, bool $force): bool
+    public function deleteCountry(int $id, bool $force = false): bool
     {
         $record = self::getById($id, withTrash: true);
+        if (!$record) {
+            bhry98_send_filament_notification("danger", __("Bhry98::notifications.filament.deleted-field"));
+            return false;
+        }
         $recordClone = self::getById($id, withTrash: true);
         if ($force) {
             $deleted = $record?->forceDelete();
             bhry98_force_delete_log((bool)$deleted, "force delete country", ['record' => $recordClone]);
+            if (!$deleted) {
+                bhry98_send_filament_notification("danger", __("Bhry98::notifications.filament.force-deleted-field"));
+            } else {
+                bhry98_send_filament_notification("success", __("Bhry98::notifications.filament.force-deleted-success"));
+            }
+
         } else {
             $deleted = $record?->delete();
             bhry98_deleted_log((bool)$deleted, "delete country", ['record' => $recordClone]);
+            if (!$deleted) {
+                bhry98_send_filament_notification("danger", __("Bhry98::notifications.filament.deleted-field"));
+            } else {
+                bhry98_send_filament_notification("success", __("Bhry98::notifications.filament.deleted-success"));
+            }
+
         }
-        return (bool)$deleted;
+         return (bool)$deleted;
     }
 
     /**
@@ -91,95 +128,96 @@ class CountriesManagementService extends BaseService
     public function restoreCountry(int $id): bool
     {
         $record = self::getById($id, withTrash: true);
+        if (!$record) {
+            bhry98_send_filament_notification("danger", __("Bhry98::notifications.filament.restored-field"));
+            return false;
+        }
         $restored = $record?->restore();
         bhry98_restored_log((bool)$restored, "restore country", ['record' => $record]);
+        if ((bool)$restored) {
+            bhry98_send_filament_notification("info", __("Bhry98::notifications.filament.restored-success"));
+        } else {
+            bhry98_send_filament_notification("danger", __("Bhry98::notifications.filament.restored-failed"));
+        }
         return (bool)$restored;
     }
 
     /**
-     * For APIs
+     * @param int $pageNumber
+     * @param int $perPage
+     * @param array|null $relations
+     * @param array|null $filters
+     * @return LengthAwarePaginator
      */
-    public function getByCode(string $code, array|null $relations = null): ?LocationsCountriesModel
-    {
-        $record = LocationsCountriesModel::query()->where('code', $code)->withCount(['governorates', 'cities', 'users']);
-        if ($relations) {
-            $record->with($relations);
-        }
-        return $record->first();
-    }
-
     public function getAll(int $pageNumber = 0, int $perPage = 20, array|null $relations = null, array|null $filters = null): LengthAwarePaginator
     {
-        $data = LocationsCountriesModel::query()
-            ->orderBy('id', 'desc');
+        $data = LocationsCountriesModel::query()->latest('id');
         if (!empty($filters)) {
             self::applyFilters($data, $filters, LocationsCountriesModel::class);
             $pageNumber = 0;
         }
-        if ($relations) {
-            $data->with($relations);
-        }
-        return $data->withCount(['users', 'governorates', 'cities'])
-            ->paginate(
-                perPage: $perPage,
-                page: $pageNumber,
-            );
+        if ($relations) $data->with($relations);
+        $data->withCount(['users', 'governorates', 'cities']);
+        return $data->paginate($perPage, page: $pageNumber);
     }
 
+    /**
+     * @param string $countryCode
+     * @param int $pageNumber
+     * @param int $perPage
+     * @param array|null $relations
+     * @param array|null $filters
+     * @return LengthAwarePaginator
+     */
     public function getAllGovernorates(string $countryCode, int $pageNumber = 0, int $perPage = 20, array|null $relations = null, array|null $filters = null): LengthAwarePaginator
     {
         $country = self::getByCode($countryCode);
-        $data = LocationsGovernoratesModel::query()
-            ->where("country_id", $country?->id ?? "")
-            ->orderBy('id', 'desc');
+        $data = LocationsGovernoratesModel::query()->where(["country_id" => $country?->id ?? ""])->latest('id');
         if (!empty($filters)) {
             self::applyFilters($data, $filters, LocationsGovernoratesModel::class);
             $pageNumber = 0;
         }
-        if ($relations) {
-            $data->with($relations);
-        }
-        return $data->withCount(['users', 'cities'])
-            ->paginate(
-                perPage: $perPage,
-                page: $pageNumber,
-            );
+        if ($relations) $data->with($relations);
+        $data->withCount(['users', 'cities']);
+        return $data->paginate($perPage, page: $pageNumber);
     }
 
+    /**
+     * @param string $countryCode
+     * @param int $pageNumber
+     * @param int $perPage
+     * @param array|null $relations
+     * @param array|null $filters
+     * @return LengthAwarePaginator
+     */
     public function getAllCities(string $countryCode, int $pageNumber = 0, int $perPage = 20, array|null $relations = null, array|null $filters = null): LengthAwarePaginator
     {
         $country = self::getByCode($countryCode);
-        $data = LocationsCitiesModel::query()
-            ->where("country_id", $country?->id ?? "")
-            ->orderBy('id', 'desc');
+        $data = LocationsCitiesModel::query()->where(["country_id" => $country?->id ?? ""])->latest('id');
         if (!empty($filters)) {
             self::applyFilters($data, $filters, LocationsCitiesModel::class);
             $pageNumber = 0;
         }
-        if ($relations) {
-            $data->with($relations);
-        }
-        return $data->withCount(['users'])
-            ->paginate(
-                perPage: $perPage,
-                page: $pageNumber,
-            );
+        if ($relations) $data->with($relations);
+        $data->withCount(['users']);
+        return $data->paginate($perPage, page: $pageNumber);
     }
 
     /**
-     * Globally
+     * @param string $countryName
+     * @param int $limit
+     * @return array
      */
     public function searchByName(string $countryName, int $limit = 20): array
     {
-        $data = LocationsCountriesModel::query();
+        $data = LocationsCountriesModel::query()->locales();
         $data->orderBy(column: 'id', direction: 'desc');
         $data->whereHas(relation: 'localizations', callback: fn($q) => $q->where('locale', app()->getLocale())->where('value', 'like', "%{$countryName}%"));
         $data->orWhereLike(column: 'default_name', value: "%{$countryName}%");
         $data->limit(value: $limit);
         $result = $data->get();
         return $result->mapWithKeys(function ($model) {
-            $label = optional($model->localizations->where('locale', app()->getLocale())->first())->value ?? $model->default_name . 33;
-            return [$model->id => $label];
+            return [$model->id => "($model->flag) $model->name"];
         })->toArray();
     }
